File structure modified and updated
