+++ models/student.go
@@ -1,0 +1,10 @@
+package models
+
+type Student struct {
+	ID    string `json:"id"`
+	Name  string `json:"name"`
+	Age   int    `json:"age"`
+	Grade string `json:"grade"`
+}
+
+var students []Student
